refactor : 데이터 클래스 분리
