+++ app/src/main/java/com/compose/hangf_aos/data/model/Customer.kt
@@ -1,0 +1,7 @@
+package com.compose.hangf_aos.data.model
+
+// 고객 정보
+data class Customer(
+    val name: String = "",
+    val phone: String = "" // Firestore에서 Document ID로 사용
+)
