+++ app/src/main/java/com/compose/hangf_aos/data/model/MenuOrder.kt
@@ -1,0 +1,9 @@
+package com.compose.hangf_aos.data.model
+
+data class MenuOrder(
+    val id: String = "",
+    val menuId: String = "",
+    val orderId: String = "",
+    val num: Int = 0,
+    val totalPrice: Int = 0
+)
