+++ app/src/main/java/com/compose/hangf_aos/data/model/OrderStatus.kt
@@ -1,0 +1,6 @@
+package com.compose.hangf_aos.data.model
+
+// 공통 타입 정의
+enum class OrderStatus {
+    Pending, Accepted, Complete, Rejected, Cancelled
+}
